register: make DataType a class, implement virtual registers
Initial registers are for alarms, sensors, active function states. One
will typically want to keep an eye on a bunch of these and having a
boatload of register nodes to read all of these in a flow is super
inconvenient to say the least.

I don't expect these to be writable however, except for maybe alarm ones
to clear the alarms. Its just hard to wrap head around it otherwise.