device: fix hang on termination
semaphore.acquire() and Promise.race() do not interact well together.
The thing is the acquire does not really get cancelled, so we just leak
a semaphore count. And there's no way to make it work with `catch` cause
the releaser gets returned by a successful acquire which `Promise.race`
basically discards.